+++ CHANGELOG.md
@@ -15,6 +15,9 @@ details.
 ### Added
 - `.eslintrc` and `npm` script command to self-lint file in this package.
 
+### Fixed
+- Lint error in .js files.
+
 ## [1.10.0] - 2017-08-10
 ### Added
 - All published rules as of `eslint-plugin-react` 7.2.0